Move server application to the 'server' package
