+++ src/main/kotlin/com/nexsabre/hardwarereservationtool/server/api/v1/Info.kt
@@ -1,4 +1,4 @@
-package com.nexsabre.hardwarereservationtool.api.v1
+package com.nexsabre.hardwarereservationtool.server.api.v1
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
